Completed the approximation of an arc as a Bézier
Also added the tests for the functions and added a bunch of nice to have
functions, like sqrt, norm, and the function to get a tangent vector.

Also created a function to convert all the vertices from one form to
another form, like form 2D to 3D for the transformation.