Rewrote the parse input function to use less memory
Also made the parse input function less clunky and ugly.

Wrote a naive implementation of Rust's str.parse function for floats
so that the program can be compiled. The parse function in Rust bloats
the data section of the binary by more than 10kB, when the
Arduino only has 8kB of memory in the data section, so this results in
a linker error and the build failing.

The function is called parse_float, and it panics if the float given
contains more than 10 digits, and returns 0.0 if the string doesn't
contain a valid float. Not exactly ideal, but is good enough for the
project.